requirements.txt
Flask
itsdangerous
requests
mctools
mysql-connector-python
gunicorn
flask-cors